src/heet/requirements.txt
astroid==2.11.5
attrs==21.4.0
autopep8==1.6.0
black==22.3.0
blessings==1.7
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
decorator==5.1.1
dill==0.3.4
earthengine-api==0.1.318
flake8==4.0.1
frictionless==4.25.1
future==0.18.2
google-api-core==2.5.0
google-api-python-client==1.12.11
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
-e git+ssh://git@github.com/UoMResearchIT/heet.git@bf5f18a19a80778bbfb9684df675e4ea5f46b63b#egg=heet
httplib2==0.20.4
httplib2shim==0.0.3
httpretty==1.1.4
idna==3.3
importlib-resources==5.4.0
iniconfig==1.1.1
isodate==0.6.1
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.4.0
lazy-object-proxy==1.7.1
marko==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.950
mypy-extensions==0.4.3
numpy==1.22.2
oauth2client==4.1.3
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pep8==1.7.1
petl==1.7.8
platformdirs==2.5.2
pluggy==1.0.0
polling2==0.5.0
prompt-toolkit==1.0.14
protobuf==3.19.4
pudb==2022.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydocstyle==6.1.1
pyfiglet==0.8.post1
pyflakes==2.4.0
Pygments==2.11.2
PyInquirer==1.0.3
pylint==2.13.9
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
pytest-socket==0.5.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
reindent==3.5.1
requests==2.27.1
rfc3986==2.0.0
rsa==4.8
shellingham==1.4.0
simpleeval==0.9.12
six==1.16.0
snowballstemmer==2.2.0
stringcase==1.2.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.62.3
transitions==0.8.11
typer==0.4.0
types-PyYAML==6.0.7
types-requests==2.27.26
types-termcolor==1.1.5
types-urllib3==1.26.14
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.8
urwid==2.1.2
urwid-readline==0.13
validators==0.18.2
wcwidth==0.2.5
wrapt==1.14.1
yaspin==2.1.0
zipp==3.7.0
